# 校园二手交易APP/openapi.yaml
openapi: 3.0.3
info:
  title: 校园二手交易 APP API
  version: 1.0.0
  description: |
    Android 客户端使用的校园二手交易平台后端接口。所有定义与 `api.md`、`db_schema.md.md`、`spec.md`、
    《后端 API 接口开发规划文档》保持一致。
servers:
  - url: http://localhost:8080/api/v1
    description: 开发环境
  - url: https://api.campus-trade.com/api/v1
    description: 生产示例
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: OK
        errorCode:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
        data:
          description: 具体数据载荷
          nullable: true
    PaginatedMeta:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        total:
          type: integer
    UserSummary:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        role:
          type: string
          enum: [STUDENT, TEACHER, STAFF]
        email: { type: string }
        phone: { type: string, nullable: true }
        avatar: { type: string, nullable: true }
        school: { type: string, nullable: true }
        emailVerified: { type: boolean }
...
    AccountStatus:
      type: object
      properties:
        deleteRequested: { type: boolean }
        deleteReason: { type: string, nullable: true }
        deleteScheduleTime: { type: string, format: date-time, nullable: true }
        status: { type: string, enum: [ACTIVE, DISABLED] }
...
    PaymentRecord:
      type: object
      properties:
        paymentId: { type: integer }
        referenceNo: { type: string }
        status:
          type: string
          enum: [PENDING, PROCESSING, SUCCEEDED, FAILED, EXPIRED, CANCELLED, REFUND_PENDING, REFUNDED]
        method:
          type: string
          enum: [ALIPAY, WECHAT, CASH]
        channel:
          type: string
          enum: [MOCK_QR, MOCK_TRANSFER, CASH]
          nullable: true
        amount: { type: number, format: double }
        currency: { type: string }
        expiresAt: { type: string, format: date-time, nullable: true }
        statusTimeline:
          type: array
          items:
            $ref: '#/components/schemas/PaymentStatusEvent'
        requestPayload:
          type: object
          nullable: true
          additionalProperties: true
        responsePayload:
          type: object
          nullable: true
          additionalProperties: true
        webhookStatus:
          type: string
          enum: [NONE, PENDING, SUCCESS, FAILED]
        callbackAt: { type: string, format: date-time, nullable: true }
        createTime: { type: string, format: date-time }
        updateTime: { type: string, format: date-time }
    PaymentStatusEvent:
      type: object
      properties:
        status:
          type: string
          enum: [PENDING, PROCESSING, SUCCEEDED, FAILED, EXPIRED, CANCELLED, REFUND_PENDING, REFUNDED]
        timestamp:
          type: string
          format: date-time
...
    BatchOperationResult:
      type: object
      properties:
        totalCount: { type: integer }
        successCount: { type: integer }
        failedCount: { type: integer }
        message:
          type: string
          nullable: true
...
    ReportResponse:
      type: object
      properties:
        id: { type: integer }
        targetType:
          type: string
          enum: [PRODUCT, USER, ORDER, MESSAGE, OTHER]
        targetId: { type: integer, nullable: true }
        targetSnapshot: { type: string, nullable: true }
        reason: { type: string }
        description: { type: string, nullable: true }
        evidenceUrls:
          type: array
          items: { type: string }
        status:
          type: string
          enum: [PENDING, IN_PROGRESS, RESOLVED, REJECTED]
        contactInfo: { type: string, nullable: true }
        resolution: { type: string, nullable: true }
        handledBy: { type: string, nullable: true }
        handledTime: { type: string, format: date-time, nullable: true }
        autoFlagged: { type: boolean }
        autoReason: { type: string, nullable: true }
        reporter:
          $ref: '#/components/schemas/UserSummary'
        createTime: { type: string, format: date-time }
    ReportDetailResponse:
      type: object
      properties:
        id: { type: integer }
        targetType:
          type: string
          enum: [PRODUCT, USER, ORDER, MESSAGE, OTHER]
        targetId: { type: integer, nullable: true }
        targetSnapshot: { type: string, nullable: true }
        reason: { type: string }
        description: { type: string, nullable: true }
        evidenceUrls:
          type: array
          items: { type: string }
        status:
          type: string
          enum: [PENDING, IN_PROGRESS, RESOLVED, REJECTED]
        contactInfo: { type: string, nullable: true }
        resolution: { type: string, nullable: true }
        handledBy: { type: string, nullable: true }
        handledTime: { type: string, format: date-time, nullable: true }
        autoFlagged: { type: boolean }
        autoReason: { type: string, nullable: true }
        createTime: { type: string, format: date-time }
        reporter:
          $ref: '#/components/schemas/UserSummary'
    ReportAuditRequest:
      type: object
      required: [status, resolution]
      properties:
        status:
          type: string
          enum: [RESOLVED, REJECTED]
        resolution:
          type: string
          maxLength: 200
    ReportBatchAuditRequest:
      type: object
      required: [ids, status, resolution]
      properties:
        ids:
          type: array
          minItems: 1
          maxItems: 200
          items: { type: integer }
        status:
          type: string
          enum: [RESOLVED, REJECTED]
        resolution:
          type: string
          maxLength: 200
...
paths:
...
  /users/password:
    put:
      summary: 修改密码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [oldPassword, newPassword]
              properties:
                oldPassword: { type: string }
                newPassword: { type: string, minLength: 6 }
      responses:
        '200': { description: 成功, content: { application/json: { schema: { $ref: '#/components/schemas/ApiResponse' } } } }
...
  /users/me/delete-request:
    post:
      summary: 申请注销
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AccountStatus'
...
  /users/me/products:
    get:
      summary: 我的商品
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [all, on_sale, off_sale, sold], default: all }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: size
          schema: { type: integer, default: 10 }
      responses:
        '200': { description: 列表, content: { application/json: { schema: { $ref: '#/components/schemas/ApiResponse' } } } }
...
  /admin/payments:
    get:
      summary: 管理员查看支付记录
      tags: [管理员]
      parameters:
        - in: query
          name: orderNo
          schema: { type: string }
        - in: query
          name: buyerPhone
          schema: { type: string }
        - in: query
          name: status
          schema:
            type: string
            enum: [PENDING, PROCESSING, SUCCEEDED, FAILED, EXPIRED, CANCELLED, REFUND_PENDING, REFUNDED]
        - in: query
          name: dateFrom
          schema: { type: string, format: date-time }
        - in: query
          name: dateTo
          schema: { type: string, format: date-time }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: size
          schema: { type: integer, default: 10 }
      responses:
        '200':
          description: 支付记录分页
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items:
                              $ref: '#/components/schemas/PaymentRecord'
                          meta:
                            $ref: '#/components/schemas/PaginatedMeta'
  /admin/payments/{paymentId}/refund:
    post:
      summary: 管理员触发退款
      tags: [管理员]
      parameters:
        - in: path
          name: paymentId
          required: true
          schema: { type: integer }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string, nullable: true }
      responses:
        '200':
          description: 退款流程已触发
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /admin/payments/{paymentId}/close:
    post:
      summary: 管理员强制关闭支付记录
      description: 将卡住的支付记录标记为 FAILED 并解除订单占用
      tags: [管理员]
      parameters:
        - in: path
          name: paymentId
          required: true
          schema: { type: integer }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string, nullable: true }
      responses:
        '200':
          description: 支付记录已关闭
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
...
  /reports/{id}:
    get:
      summary: 举报详情
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 举报详细信息
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ReportDetailResponse'
...
  /admin/moderation/reports:
    get:
      summary: 举报列表
      description: 管理员按条件筛选举报记录，默认按自动预警优先 + 创建时间倒序
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [PENDING, IN_PROGRESS, RESOLVED, REJECTED] }
        - in: query
          name: targetType
          schema: { type: string, enum: [PRODUCT, USER, ORDER, MESSAGE, OTHER] }
        - in: query
          name: reporterKeyword
          schema: { type: string }
        - in: query
          name: autoFlaggedOnly
          schema: { type: boolean }
        - in: query
          name: startTime
          schema: { type: string, format: date-time }
        - in: query
          name: endTime
          schema: { type: string, format: date-time }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: size
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        '200':
          description: 举报列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items: { $ref: '#/components/schemas/ReportResponse' }
                          meta:
                            $ref: '#/components/schemas/PaginatedMeta'
...
  /admin/moderation/reports/{id}:
    get:
      summary: 举报详情
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 举报详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ReportDetailResponse'
    patch:
      summary: 处理举报
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportAuditRequest'
      responses:
        '200':
          description: 举报已处理
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /admin/moderation/reports/batch:
    post:
      summary: 批量处理举报
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportBatchAuditRequest'
      responses:
        '200':
          description: 批量处理结果
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BatchOperationResult'
...